Added all region lists
Also my plotting fractional coverage.ipynb had a warning on a change that I didn't remember making stating "Warning: line endings have changed from "LF" to "CRLF"".  I'm unsure what this entails but figure I can push it and if it was something we didn't want we can revert it to an earlier state.